k8s-flask-service/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.9
kubernetes==32.0.1
flask-swagger-ui==4.11.1
python-dotenv==1.0.1